requirements.txt
asgiref==3.8.1
Django==5.1.1
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework-simplejwt==5.3.1
gis==0.2.1
Markdown==3.7
osgeo4w_installer==1.4.0
pillow==10.4.0
postgres==4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
psycopg2-pool==1.2
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.2
